backend/requirements.txt
async-timeout==4.0.3
faiss-cpu==1.8.0
fastapi==0.111.0
# fastapi-cli==0.0.4
httpx==0.27.0
# huggingface-hub==0.17.3
# langchain==0.2.5
# langchain-community==0.2.5
# langchain-core==0.2.7
# langchain-text-splitters==0.2.1
numpy==1.26.4
orjson==3.10.5
pandas==2.2.2
sentence-transformers==2.7.0
# tokenizers==0.14.1
torch==2.3.0
transformers==4.34.1
underthesea==6.8.0
underthesea_core==1.0.4